feat(add-article-auto-generate): display published date for news and modify generate codes.
